+++ inc/staffSvg/organizeTokens.ts
@@ -377,17 +377,22 @@ const parseTokenStaff = ({tokens, y, top, measureRanges, logger}) => {
 
 		const nearest = notes.reduce((best, note) => {
 			if (!note.tied) {
-				const dx = note.x + constants.NOTE_TYPE_WIDTHS[note.noteType] * 0.5 - position.x;
-				if (dx > -1) {
-					const dy = (note.y - position.y) * 2;
-					const distance = Math.sqrt(dx * dx + dy * dy);
-					if (distance < best.distance)
-						return {distance, note};
+				const distance = note.sourceProgress - tie.sourceProgress;
+				if (distance > 0 && distance < best.distance) {
+					const dx = note.x + constants.NOTE_TYPE_WIDTHS[note.noteType] * 0.5 - position.x;
+					if (dx > -1) {
+						const dy = (note.y - position.y) * 2;
+						const tipDistance = Math.sqrt(dx * dx + dy * dy);
+						if (tipDistance < 8) {
+							//logger.append("nearNote", {tipDistance, tie: tie.sourceProgress, note: note.sourceProgress});
+							return {distance, note};
+						}
+					}
 				}
 			}
 
 			return best;
-		}, {distance: 8});
+		}, {distance: Infinity});
 		if (nearest.note) {
 			nearest.note.tied = true;
 			logger.append("parseTokenStaff.tiedNote", {nearest, tie});