+++ backend/scoreMaker.ts
@@ -74,7 +74,7 @@ const xmlBufferToLy = async (xml: Buffer, options: LilyProcessOptions = {}): Pro
 const unescapeStringExp = exp => exp && exp.toString();
 
 
-const markScoreV1 = async (source: string, lilyParser: GrammarParser, {midi, logger}: {midi?: MIDI.MidiData, logger?: LogRecorder} = {}): Promise<ScoreJSON> => {
+const makeScoreV1 = async (source: string, lilyParser: GrammarParser, {midi, logger}: {midi?: MIDI.MidiData, logger?: LogRecorder} = {}): Promise<ScoreJSON> => {
 	const t0 = Date.now();
 
 	const engraving = await engraveSvg(source);
@@ -116,7 +116,7 @@ const markScoreV1 = async (source: string, lilyParser: GrammarParser, {midi, log
 		const coverage = ((matcher.criterion.notes.length - omitC) / matcher.criterion.notes.length)
 			* ((matcher.sample.notes.length - omitS) / matcher.sample.notes.length);
 
-		logger.append("markScore.match", {coverage, omitC, omitS, path: matcher.path});
+		logger.append("makeScore.match", {coverage, omitC, omitS, path: matcher.path});
 	}
 
 	const matchedIds: Set<string> = new Set();
@@ -151,7 +151,13 @@ interface IncompleteScoreJSON {
 };
 
 
-const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, logger}: {midi?: MIDI.MidiData, logger?: LogRecorder} = {}): Promise<ScoreJSON | IncompleteScoreJSON> => {
+interface SheetNotationResult extends IncompleteScoreJSON {
+	midiNotation: MusicNotation.NotationData;
+	sheetNotation: staffSvg.StaffNotation.SheetNotation;
+};
+
+
+const makeSheetNotation = async (source: string, lilyParser: GrammarParser, {withNotation = false, logger}: {withNotation?: boolean, logger?: LogRecorder} = {}): Promise<SheetNotationResult> => {
 	let midiNotation = null;
 
 	const pages = [];
@@ -169,14 +175,11 @@ const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, log
 			attrGen.release(lilyDocument.globalAttributes({readonly: true}) as LilyDocumentAttributeReadOnly);
 			//console.log("tp.1:", Date.now() - t0);
 		},
-		onMidiRead: midi_ => {
+		onMidiRead: withNotation && (midi => {
 			//console.log("tm.0:", Date.now() - t0);
-			if (!midi) {
-				midi = midi_;
-				midiNotation = midi && MusicNotation.Notation.parseMidi(midi);
-			}
+			midiNotation = midi && MusicNotation.Notation.parseMidi(midi);
 			//console.log("tm.1:", Date.now() - t0);
-		},
+		}),
 		onSvgRead: async svg => {
 			//console.log("ts.0:", Date.now() - t0);
 			const attributes = await attrGen.wait();
@@ -187,14 +190,10 @@ const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, log
 		},
 	});
 
-	//console.log("t2:", Date.now() - t0);
-
 	logger.append("scoreMaker.profile.engraving", {cost: Date.now() - t0});
 	logger.append("lilypond.log", engraving.logs);
 
 	const doc = new staffSvg.SheetDocument({pages});
-
-	//console.log("t3:", Date.now() - t0);
 
 	const attributes = await attrGen.wait();
 	const meta = {
@@ -204,6 +203,23 @@ const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, log
 		pageCount: doc.pages.length,
 		staffSize: attributes.staffSize,
 	};
+
+	const sheetNotation = staffSvg.StaffNotation.parseNotationFromSheetDocument(doc, {logger});
+
+	return {
+		midiNotation,
+		sheetNotation,
+		meta,
+		doc,
+		hashTable,
+	};
+};
+
+
+const makeScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, logger, unfoldRepeats = false}: {midi?: MIDI.MidiData, logger?: LogRecorder, unfoldRepeats?: boolean} = {}): Promise<ScoreJSON | IncompleteScoreJSON> => {
+	const t0 = Date.now();
+
+	const {midiNotation, sheetNotation, meta, doc, hashTable} = await makeSheetNotation(source, lilyParser, {logger, withNotation: true});
 
 	if (!midiNotation) {
 		console.warn("Neither lilypond or external arguments did not offer MIDI data, score maker finish incompletely.");
@@ -215,19 +231,13 @@ const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, log
 		}; 
 	}
 
-	//console.log("t4:", Date.now() - t0);
-
-	const sheetNotation = staffSvg.StaffNotation.parseNotationFromSheetDocument(doc, {logger});
-
-	//console.log("t5:", Date.now() - t0);
 	const t5 = Date.now();
 
 	const matcher = await staffSvg.StaffNotation.matchNotations(midiNotation, sheetNotation);
 
-	//console.log("t6:", Date.now() - t0);
 	logger.append("scoreMaker.profile.matching", {cost: Date.now() - t5});
 
-	if (logger) {
+	if (logger && logger.enabled) {
 		const cis = new Set(Array(matcher.criterion.notes.length).keys());
 		matcher.path.forEach(ci => cis.delete(ci));
 
@@ -237,27 +247,18 @@ const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, log
 		const coverage = ((matcher.criterion.notes.length - omitC) / matcher.criterion.notes.length)
 			* ((matcher.sample.notes.length - omitS) / matcher.sample.notes.length);
 
-		logger.append("markScore.match", {coverage, omitC, omitS, path: matcher.path});
+		logger.append("makeScore.match", {coverage, omitC, omitS, path: matcher.path});
 	}
-
-	//console.log("t7:", Date.now() - t0);
 
 	const matchedIds: Set<string> = new Set();
 	midiNotation.notes.forEach(note => note.ids && note.ids.forEach(id => matchedIds.add(id)));
 
-	//console.log("t8:", Date.now() - t0);
-
 	doc.updateMatchedTokens(matchedIds);
-
-	//console.log("t9:", Date.now() - t0);
 
 	const pitchContextGroup = staffSvg.StaffNotation.createPitchContextGroup(sheetNotation.pitchContexts, midiNotation);
 
-	//console.log("t10:", Date.now() - t0);
-
 	const noteLinkings = midiNotation.notes.map(note => _.pick(note, ["ids", "staffTrack", "contextIndex"]));
 
-	//console.log("t11:", Date.now() - t0);
 	logger.append("scoreMaker.profile.full", {cost: Date.now() - t0});
 
 	return {
@@ -271,14 +272,14 @@ const markScoreV2 = async (source: string, lilyParser: GrammarParser, {midi, log
 };
 
 
-const markScore = markScoreV2;
+const makeScore = makeScoreV2;
 
 
 
 export {
 	markupLily,
 	xmlBufferToLy,
-	markScore,
-	markScoreV1,
-	markScoreV2,
+	makeScore,
+	makeScoreV1,
+	makeScoreV2,
 };