+++ tests/bundleScore.ts
@@ -2,7 +2,7 @@
 import fs from "fs";
 
 import "../env.js";
-import {markScoreV1, markScore} from "../backend/scoreMaker";
+import {makeScore} from "../backend/scoreMaker";
 import loadLilyParser from "../backend/loadLilyParserNode";
 import LogRecorder from "../inc/logRecorder";
 
@@ -15,7 +15,7 @@ const main = async (...args) => {
 
 	const ly = fs.readFileSync(args[0]).toString();
 	const t0 = Date.now();
-	const result = await markScore(ly, lilyParser, {logger});
+	const result = await makeScore(ly, lilyParser, {logger});
 	console.log("cost:", Date.now() - t0);
 	//console.log("result:", result, logger);
 	console.log("profiles:", logger.records.filter(r => /scoreMaker\.profile/.test(r.desc)));